Combine all snapcast views into one file.
It's a little bigger than I'd like right now, but I can revisit it.
