+++ vulpes/blueprints/snapcast/sql.py
@@ -1,5 +1,4 @@
 from datetime import datetime, timezone
-from uuid import UUID
 
 from sqlalchemy import update
 
@@ -7,7 +6,7 @@ from ... import db
 from .models import Podcast
 
 
-def touch_podcast(podcast_uuid: UUID):
+def touch_podcast(podcast_uuid):
     """Update the last_modified field for a podcast. Called on cache-invalidating requests."""
     db.session.execute(
         update(Podcast)
